Refactor common into single file
