--- alignment-software/alignment_software/gui/numericspinbox.py
@@ -1,79 +1,0 @@
-from tkinter import ttk
-
-
-class NumericSpinbox(ttk.Spinbox):
-    """A spinbox with a validated numeric range and type."""
-
-    def __init__(
-        self, master, value_default=0, value_range=[0, 100], value_type=int,
-        command=None, **kwargs
-    ):
-        """Creates a numeric spinbox."""
-        super().__init__(
-            master, from_=value_range[0], to=value_range[1], **kwargs
-        )
-        self.value_type = value_type
-        self.value_range = value_range
-        self.value_default = value_default
-        self.value_cached = None
-        self.command = command
-        validate_command = (master.register(self.validate), '%P')
-        invalid_command = (master.register(self.on_invalid),)
-        self.config(validate="focusout", validatecommand=validate_command)
-        self.config(invalidcommand=invalid_command, command=self.on_change)
-        self.set(value_default)
-
-    def get(self):
-        """Returns the numeric value of the spinbox."""
-        return self.value_type(super().get())
-
-    def set(self, value):
-        """Sets the numeric value of the spinbox."""
-        self.value_cached = value
-        super().set(value)
-
-    def set_command(self, command):
-        """Sets the command to be called when the spinbox in updated."""
-        self.command = command
-
-    def set_value_range(self, minimum, maxmimum):
-        """Updates the range of values allowed in the spinbox."""
-        self.config(from_=minimum, to=maxmimum)
-        self.value_range = (minimum, maxmimum)
-        if not self.validate(self.value_cached):
-            self.on_invalid()
-
-    def on_change(self):
-        """
-        Runs on any value change.
-        Calls validate, but does not call on_invalid to correct value.
-        """
-        self.validate(self.get())
-
-    def validate(self, value):
-        """
-        Returns validaty of the spinbox.
-        Calls the command if the value is updated and it is valid.
-        """
-        try:
-            value = self.value_type(value)
-            if self.value_range is not None:
-                if value < self.value_range[0]:
-                    valid = False
-                if value > self.value_range[1]:
-                    valid = False
-            valid = True
-        except Exception:
-            valid = False
-        if valid and self.value_cached != value:
-            self.value_cached = value
-            if self.command is not None:
-                self.command()
-        return valid
-
-    def on_invalid(self):
-        """Updates the value to a previous valid value when invalid."""
-        if self.value_cached is not None:
-            self.set(self.value_cached)
-        else:
-            self.set(self.value_default)
